backend/requirements.txt
-i https://pypi.org/simple
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
deprecated==1.2.13
flower==1.0.0
humanize==4.1.0
kombu==5.2.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pyparsing==3.0.9
redis==4.3.1
six==1.16.0
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
protobuf==4.23.4
psycopg2-binary==2.9.9
pytz==2023.3
pywatchman==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
whitenoise==6.6.0
djangorestframework-simplejwt
django-allauth
dj-rest-auth
drf-nested-routers
setuptools==62.3.2
urllib3==2.0.6
zipp==3.17.0
Pillow==10.1.0
mysqlclient==2.1.0
idna==3.4
django-environ==0.6.0
